Move views.py to its own folder.
